Move Authentication class to ebrains module, new siibra.set_ebrains_token() function
